refactor models to use abstract data types; include a simple `Midi-to-Json` converter
